rettelser for det hele, + SaleItem og Sale post og get kan laves

+++ src/main/java/com/example/kinoxpbackend/repositoryAllSale/SaleDetailRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SaleDetailRepository extends JpaRepository<SaleDetail, Integer> {
+public interface SaleDetailRepository extends JpaRepository<SaleDetail, Long> {
 }
